+++ tests/testthat/test-validate.R
@@ -2,36 +2,42 @@ testthat::context("Test round trip validation")
 
 library(xml2)
 
-guess_ns <- function(file){
-  x <- xml2::xml_root(xml2::read_xml(file))
-  root <- xml2::xml_name(x)
-  ns <- xml2::xml_ns(x)
-  i <- grep(strsplit(xml2::xml_attr(x, "schemaLocation"), "\\s+")[[1]][1], ns)
-  ns <- names(ns[i])[[1]]
-  list(root=root, ns=ns)
-}
-
-
 test_roundtrip <- function(f, schema = NULL, check_lengths = TRUE){
   testthat::test_that(paste(
     "testing that", basename(f), "can roundtrip & validate"),
   {
 
   ## guess root and ns for sub-modules
-  ns <- guess_ns(f)
+  doc <- xml2::read_xml(f)
+  ns <- find_real_root_name(doc)
 
   out <- tempfile(basename(f), fileext = ".xml")
 
   emld <- as_emld(f)
-  as_xml(emld, out, ns$root, ns$ns)
+  as_xml(emld, out, ns$name, ns$prefix)
 
-  ## Make sure output xml is still valid
-  testthat::expect_true( eml_validate(out, schema = schema) )
+  ## Make sure output xml is still valid unless it's supposed to be invalid
+  if (!grepl("invalid", f, ignore.case = TRUE)) {
+    testthat::expect_true( eml_validate(out, schema = schema) )
+
+  } else {
+    testthat::expect_false( eml_validate(out, schema = schema) )
+  }
 
   ## Make sure we have the same number & names of elements as we started with
   if(check_lengths){
     elements_at_end <- sort(names(unlist(as_emld(out), recursive = TRUE)))
     elements_at_start <- sort(names(unlist(emld, recursive = TRUE)))
+
+    # Filter out schemaLocation since it can be absent at start and added at end
+    elements_at_start <- Filter(
+      function (e) { e != "schemaLocation"},
+      elements_at_start)
+
+    elements_at_end <- Filter(
+      function (e) { e != "schemaLocation"},
+      elements_at_end)
+
     testthat::expect_equal(elements_at_start, elements_at_end)
   }
   })
@@ -155,3 +161,80 @@ partial_test <- basename(suite) %in%
     "eml-literatureInPress.xml",
     "eml-citationWithContact.xml")
 lapply(suite[partial_test], test_roundtrip, check_lengths = FALSE)
+
+
+######## Test that new schema validation logic that doesn't use schemaLocation
+######## still validates
+test_that("we can still validate XML docs without schemaLocation set on them", {
+  testthat::expect_true(
+    eml_validate(
+      system.file(
+        file.path("tests",
+                  "eml-2.2.0",
+                  "eml-datasetNoSchemaLocation.xml"),
+        package="emld")
+    ))
+
+  testthat::expect_false(
+    eml_validate(
+      system.file(
+        file.path("tests",
+                  "eml-2.2.0",
+                  "eml-datasetNoSchemaLocationInvalid.xml"),
+        package="emld")
+    ))
+})
+
+
+######### Helper methods
+test_that("get_root_ns works for a variety of cases", {
+  testthat::expect_equal(
+    find_real_root_name(
+      xml2::read_xml("<dictionary
+                        xmlns=\"http://www.xml-cml.org/schema/stmml-1.1\">
+                     </dictionary>")),
+    list(prefix = "d1", name = "dictionary"))
+
+  testthat::expect_equal(
+    find_real_root_name(
+      xml2::read_xml("<eml:eml
+                        xmlns:eml=\"https://eml.ecoinformatics.org/eml-2.2.0\">
+                     </eml:eml>")),
+    list(prefix = "eml", name = "eml"))
+
+  testthat::expect_equal(
+    find_real_root_name(
+      xml2::read_xml("<cit:citation
+                        xmlns:cit=\"https://eml.ecoinformatics.org/literature-2.2.0\">
+                     </cit:citation>")),
+    list(prefix = "cit", name = "citation"))
+})
+
+test_that("get_root_ns works for a variety of cases", {
+  testthat::expect_equal(
+    guess_root_schema(
+      xml2::read_xml("<dictionary
+                        xmlns=\"http://www.xml-cml.org/schema/stmml-1.1\">
+                     </dictionary>")),
+    list(module = "stmml",
+         version = "1.1",
+         namespace = "http://www.xml-cml.org/schema/stmml-1.1"))
+
+  testthat::expect_equal(
+    guess_root_schema(
+      xml2::read_xml("<eml:eml
+                        xmlns:eml=\"https://eml.ecoinformatics.org/eml-2.2.0\">
+                     </eml:eml>")),
+    list(module = "eml",
+         version = "2.2.0",
+         namespace = "https://eml.ecoinformatics.org/eml-2.2.0"))
+
+  testthat::expect_equal(
+    guess_root_schema(
+      xml2::read_xml("<cit:citation
+                        xmlns:cit=\"https://eml.ecoinformatics.org/literature-2.2.0\">
+                     </cit:citation>")),
+    list(module = "literature",
+         version = "2.2.0",
+         namespace = "https://eml.ecoinformatics.org/literature-2.2.0"))
+})